+++ CHANGELOG.md
@@ -1,8 +1,22 @@
+## v4.0.0
+
+- New✨ - Now we can add our own custom apps to generate a interactive ui.
+- Added - `Vite` and `Vite Extra` App
+- Renamed - `create.app` command to `create-app.interactive` command.
+  - Give `(cmd/ctrl)+shit+p` and select `Create App: Interactive` to open the interactive ui.
+- Added - `create-app.quick` command.
+  - Give `(cmd/ctrl)+shit+p` and select `Create App: Quick` to open the quick command pallet to create a app
+- Added - `create-app.settings.customApps` setting - provide a direct list of custom apps config
+- Added - `create-app.settings.customAppPath` setting - provide the `json` file path or folder path containing list of custom apps config.
+- Added - `copy config` button ont he top right corner of the command text area. Helps to copy the app configurations and override the default configs.
+- Removed - `Open in vscode` config from all app configs.
+- Fixed - Keeps on reloading on view switching the tabs issue fixed.
+
 ## v3.1.0
 
 - Typo fixes.
 - Build size reduced.
-  
+
 ## v3.0.2
 
 - Due to compatibility reverted back [v3.0.1](#v3.0.1)
@@ -11,7 +25,7 @@
 
 - On `Execute` button click, the terminal opens in powershell and failed to execute the create app commands - `Fixed`
 - Now The Terminal open in `Git bash` by default on click of the `Execute` button.
-  
+
 ## v3.0.0
 
 - Added Gatsby App